project/requirements.txt
fastapi
uvicorn
opencv-python
numpy
pandas
requests
pillow
matplotlib
scikit-learn
websockets
python-multipart
passlib[bcrypt]
pyjwt
sqlalchemy
fpdf2
python-dotenv
pymongo[srv]
